refactor(network): move RayonAsyncComputationSpawner to near_async
Address review1.md T1 B1: deduplicate `RayonAsyncComputationSpawner`
that previously lived as separate copies in `chain/chain/src/rayon_spawner.rs`
and `chain/network/src/concurrency/rayon.rs`. The chain/chain copy
preserved the tracing dispatcher across the rayon hop; the chain/network
copy did not (silently dropped span context for signature verification).

The crates can't share via re-export (chain/chain depends on chain/network,
so chain/network can't depend on chain/chain). Move the canonical impl
to `near_async` (which both depend on transitively):

- `core/async/src/rayon_spawner.rs`: canonical `RayonAsyncComputationSpawner`
  with tracing dispatcher propagation.
- `core/async/Cargo.toml`: add `rayon` dependency.
- `core/async/src/lib.rs`: `pub use rayon_spawner::RayonAsyncComputationSpawner`.
- `chain/network/src/concurrency/rayon.rs`: `pub use near_async::RayonAsyncComputationSpawner`.
- `chain/chain/src/rayon_spawner.rs`: `pub use near_async::RayonAsyncComputationSpawner`.

All ~39 existing call sites continue to compile unchanged through the
re-exports. Behavior matches `chain/chain`'s previous impl (tracing
dispatcher propagated).

Also addresses review1.md T1 S1: PMA spawner field doc updated to
reflect the post-T0 contract — the spawner's runtime lifetime is the
caller's responsibility; spawned tasks are best-effort.
