+++ core/async/Cargo.toml
@@ -16,6 +16,7 @@ crossbeam-channel.workspace = true
 futures.workspace = true
 parking_lot.workspace = true
 rand = { workspace = true, optional = true }
+rayon.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 time.workspace = true
